+++ algo_training/two_towers_training/tools/utils.py
@@ -1,12 +1,38 @@
+import os
+
+import mlflow
 import tensorflow as tf
 import pandas as pd
 
-import shutil
 from loguru import logger
+
+from google.auth.transport.requests import Request
+from google.cloud import secretmanager
+from google.oauth2 import id_token
+
+from tools.constants import GCP_PROJECT_ID, MLFLOW_EHP_URI, MLFLOW_PROD_URI
 
 from sklearn.decomposition import PCA
 import matplotlib as mpl
 import matplotlib.pyplot as plt
+
+
+def get_secret(secret_id: str):
+    client = secretmanager.SecretManagerServiceClient()
+    name = f"projects/{GCP_PROJECT_ID}/secrets/{secret_id}/versions/1"
+    response = client.access_secret_version(name=name)
+    return response.payload.data.decode("UTF-8")
+
+
+def connect_remote_mlflow(client_id, env="ehp"):
+    """
+    Use this function to connect to the mlflow remote server.
+
+    client_id : the oauth iap client id (in 1password)
+    """
+    os.environ["MLFLOW_TRACKING_TOKEN"] = id_token.fetch_id_token(Request(), client_id)
+    uri = MLFLOW_PROD_URI if env == "prod" else MLFLOW_EHP_URI
+    mlflow.set_tracking_uri(uri)
 
 
 def save_pca_representation(
